executive-ops-copilot-v0/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "executive-ops-copilot-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "uvicorn[standard]>=0.29",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-mock>=3.14",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
